Refactoring the schema diff/validation -> core
