--- builtin/providers/azurerm/structure.go
@@ -1,24 +1,0 @@
-package azurerm
-
-import "encoding/json"
-
-// Takes a value containing JSON string and passes it through
-// the JSON parser to normalize it, returns either a parsing
-// error or normalized JSON string.
-func normalizeJsonString(jsonString interface{}) (string, error) {
-	var j interface{}
-
-	if jsonString == nil || jsonString.(string) == "" {
-		return "", nil
-	}
-
-	s := jsonString.(string)
-
-	err := json.Unmarshal([]byte(s), &j)
-	if err != nil {
-		return s, err
-	}
-
-	bytes, _ := json.Marshal(j)
-	return string(bytes[:]), nil
-}
